GPS: Add "Add multi-cursor to all references of entity" command to the multicursors python plugin.
(multicursors.py) : Add said command

(Src_Editor_Module.adb): Fix an ugly bug which causes any EditorLocation
created on the python side with the default constructor to have invalid
properties on the Ada side (because of lack of initialization).
The fix is basically to use the already existing Create_Editor_Location
constructor on the Ada side, which uses iterators and intialize all
fields.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/branches/global/gtk3/gps@210734 936e1b1b-40f2-da11-902a-00137254ae57