* gvd-preferences.adb: (Register_Default_Preferences): Prefix "Debugger-Kind" with "GPS-6" since the  value of the enum has changed during the work on gdb-mi.
* default_preferences.adb:
(Load_Preferences): Fix logic error introduced in previous checkin.
(Add_Theme): Add safety against adding a theme multiple times. Do not set the
 current theme to a "Dark" variant by default.
(Create): Hard-code the presence of the Adwaita theme under Windows.

* gps-kernel-preferences.adb:
Add the "GPS6-" prefix to "Gtk-Theme-Name", forcing the switch to Adwaita.

* colorschemes.py:
Adapt to the new name of Gtk-Theme-Name

* icons.py:
Adapt to the new name of "Gtk-Theme-Name".

For M731-015.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@215782 936e1b1b-40f2-da11-902a-00137254ae57