Fix for GPS 6.0 : Drag and drop messes up actions queue. M906-024
* (Src_Editor_View): Add a callback for the "drag-data-received" signal.
In this callback, test if the target is GTK_TEXT_BUFFER_CONTENTS. In
this case, we know that Gtk is attempting to "smart copy" the content of
a Gtk Text Buffer to another Gtk Text Buffer. In this case, block the
default behavior and implement ad hoc dropping of text.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@215874 936e1b1b-40f2-da11-902a-00137254ae57