(Register_Action): can be called without creating temporary variable.
It is now possible to call Register_Action and directly call
"new command" in the list of parameters, rather than create an
intermediate local variable. This is a minor cleanup, but removes
a number of lines in the GPS source code.

(Register_Action): removed function version, since unused

Remove the Kernel field from all child types of Interactive_Command,
since the kernel is also available through the context. Minor simplification.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@219645 936e1b1b-40f2-da11-902a-00137254ae57