Initial support for the Google Closure compiler
This check in provides a compile.sh script which will download if needed
the compiler, then run it (and the style checker) on our javascript
sources.
It also adds a number of type annotatations to the sources as an example,
already reducing significantly the number of errors, but not fixing all
of them.
Introduces a GNATdoc namespace, which the *.js.tmpl file are now using.
This provides a way to describe the expected type for the JSON data, so
that the compiler can do its work.

Part of N109-062

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@222307 936e1b1b-40f2-da11-902a-00137254ae57