Fix loading color preferences declared in python plug-ins.
When a color preference is declared in a python plug-in, it might be created
after the preferences file has been loaded. GPS then did not properly
convert that preference from a string type to a color type preference, and
would always use the default value set through GPS.Preference.create as
the current value.

In particular, this was the case for the ispell.py plug-in.

N214-038

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@223406 936e1b1b-40f2-da11-902a-00137254ae57