Add support for specifing object file extension for languages
The goal is to be able to set it to "-" for languages that have no
object files. This is used to avoid naming collisions when
GNATCOLL.Projects.Library_Files finds a ".*li" file and tries to
find which project it belongs to. For instance, the plugin unicode.py
had the result that the file unicode.ali in the xmlada install was
though to belong to the plugins.gpr project, and thus unicode.ali
was never returned, and thus its cross references never parsed.

This object file suffix can be set either in GPS plugins like
languages.py, or explicitly in project files, which was done here
so that gnatinspect is not confused.

N304-018

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224037 936e1b1b-40f2-da11-902a-00137254ae57