Add support for showing warnings in console when sqlite database is corrupted
In GNATCOLL it is now possible to register an error handler which has
several primitive operations that will be called when a warning or an
error is detected, and also when the database is corrupted.
We take advantage of this in GPS to display a warning in the console
when corruption is detected.
N224-020

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224074 936e1b1b-40f2-da11-902a-00137254ae57