"Compile File" should use the file's project to recompile, not root project
This provides a better behavior in the case of aggregate projects,
since otherwise gprbuild seems to be recompiling more than just the
file itself.
This was not so easy before because the source editors did not
provide the project information in the context, but this has been
changed now.

N305-035

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224396 936e1b1b-40f2-da11-902a-00137254ae57