Fix name of entity when searching for all references.
By default, we also look for references to renamed entities, but we
were always displaying the name of the renaming entity.
Fix order of declarations in xref.ads to avoid freezing issues. For
instance it was not possible to use dot-notation to call Get_Name
on a General_Entity.
(Get_Entity): did not properly initialize the db field, so it was not
possible to do a Get_Entity (ref).Get_Name

part of N319-019

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@224736 936e1b1b-40f2-da11-902a-00137254ae57