N407-023: new highlighting framework for GPS.
This is the first commit for the implementation of a new highlighting
framework for GPS, which is described in N407-023. Here are the key
design points, as described in the TN:

1. Incremental and generator based (can be stopped and resumed)
2. Supports clever rehighlighting of portions of a buffer, so tends to
exhibit better performance on complex regions than the current
highlighter (for multiline comments for example)
3. Allows the client to specify syntaxes in a declarative way. Extending
GPS to support a new language is made much easier (and much more
powerful than the basic keyword capability we had before).
4. Uses regexes for tokenization.
5. Separation between the language specification and the engine core,
makes it easier to understand and maintain.
6. Pure python.

This commit contains the core engine, as well as an highlighter for C
and an highlighter for python. It deactivates the current highlighters
for C and python, but the current highlighter for C++ stays, since
support for this language is not complete.

The highlighter framework is already quite well commented, but further
work is needed to document it completely.

File changes:

* Language.C (Get_Language_Context): Deactivate built-in syntax
  highlighting for C.

* share/plug-ins/python_support.py (Python_Support): Deactivate built-in
  syntax highlighting for Python.

* share/support/ui/highlight_framework.py: Core engine responsible for
  the highlighting.

* share/support/ui/highlight_c.py: Definition of the C highlighter.

* share/support/ui/highlight_py.py: Definition of the python
  highlighter.

* design/test/GPS_Support.py (compare_to_file): New test support
  function for comparing the content of a buffer to a file.

* design/test/multicursors/test.py: Remove compare to file fn

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@225740 936e1b1b-40f2-da11-902a-00137254ae57