Add support for saving and reloading a Simulink file as part of the desktop
This required various modifications in GPS, because the custom browsers are
using an instance of generic_views.ads (so that we can have local toolbars)
but using a python module's save_desktop and load_desktop function requires
that GPS.MDI.add be called with a proper save_desktop callback.
As such setting such a callback is now possible directly on a GPS_MDI_Child
rather than on a hidden class in python_module.adb.

Saving the desktop will now save the name of the .mdl file, and automatically
reload it when GPS is reopened.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@225868 936e1b1b-40f2-da11-902a-00137254ae57