* src_editor_module.adb: (Project_View_Changed): New subprogram. (Register_Module): Connect Project_View_Changed.
* src_editor_view.adb:
(Initialize): Factor code in Update_Project.
(Update_Project): New subprogram.

* src_editor_view.ads:
(Update_Project): New subprogram, to recompute the Project cached in the
 editor view.

With the introduction of aggregate projects, we cache the project in editor
view. This causes dangling references when the project changes but the
views stay open - for instance when using the GNATtest integration to switch
to the harness project.

To solve this, react to project_view_changed: walk all open editors and
recompute the project cache then.

Tentative fix for N411-039




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@226408 936e1b1b-40f2-da11-902a-00137254ae57