Redo patch for N411-039
Previous patch (refreshing the cache of project in editors when project
is reloaded) did not quite work: on project reload, the editor would fall
back to the first possible project containing this file, disregarding
the fact that a specific other project might have been chosen before.
So now we cache the project path instead of the project_type.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@226556 936e1b1b-40f2-da11-902a-00137254ae57