GPS: Big refactoring of the xref engine interface, meant to finish the job started by Nico for N301-004. The first part of the job was to :
- Make Entity an abstract tagged type, and make every operation on
entities a dispatching operation on entities.
- In this scenario, the database serves as a factory/dispatcher, that
will return an entity of the appropriate concrete type depending on the
context (file/language/etc).

This job was not finished, because the Entity type was made abstract,
but not entity references and iterators. The remaining work was thus the
following:

1. Make Entity_Iterator and Entity_Reference abstract tagged types too,
and owning a reference to their database, so as to eliminate Db params,
and make all operations dispatching on one of those.
2. Refactor all the client code to take the modifications into account,
mainly using holders for types of indefinite sizes.

A trick was used for 2., declaring the the root abstract types as public
null records so that we can declare holders to them in the Kernel.Xref
package. This commit gets rid of the multiple Indefinite_Holder
instantiations made by the previous commit.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@227393 936e1b1b-40f2-da11-902a-00137254ae57