Add support for monitor files from any view
The mechanism used in editors to tell users that a file has changed on
disk and the editor needs to be reloaded has been moved to the kernel,
so that it can be used in other types of views.
A single dialog is now displayed when any number of file has changed on
the disk, not one per file, which makes it more user-friendly.

Part of N513-036 and LA18-019

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@227770 936e1b1b-40f2-da11-902a-00137254ae57