Add support for key themes
GPS comes with two key themes (default and Emacs). The emacs key theme replaces
 the emacs.py plug-in, and should be easier to find for users.
Switching key themes dynamically is possible via the key shortcuts dialog (which
 was not possible with the emacs plug-in), and GPS will take care of removing the
 shortcuts coming from the previous scheme (only preserving the ones set manually
 by the user).
Add support for "cmd-" as a key modifier in the XML files that describe the key
 themes. Update documentation accordingly.
Move isearch.py to plug-ins/ to load it by default. This was previously done by
 emacs.py, and it is better for users looking for Emacs compatibility. This
 plug-in is also tested enough that we should have it by default now.
Ensure the shortcuts displayed in the menus are always correct with the current
 theme. This fixes a long-standing issue in GPS where the menus were not
 properly updated after key bindings were changed or removed, which became
 even more visible when the user changes the theme dynamically.
Add support for creating user themes
M917-006

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@228198 936e1b1b-40f2-da11-902a-00137254ae57