Add support for showing actions with no category in Key Shortcuts
In addition, move all the help actions to no-category, so that they
are not displayed by default. This keeps the list of actions shorter
by default.
(On_Create): make sure the dialog is a child of the Key Shortcuts
dialog, not of the GPS main window, so that it doesn't go below the
dialog.
M917-006

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@228356 936e1b1b-40f2-da11-902a-00137254ae57