(Find_Utils): reimplemented with GPS.Search
This simplifies the code (avoids duplication), and makes it possible to over
search types (approximative or fuzzy) in the future for the search contexts
(like Project view).
(Update_Location): fixes some issues where the current line and column were
counted incorrectly (which was not very visible in the omni-search, but much
more in the editor).
Implement support for Whole_Word in GPS.Search (which had it in the API, but
no implementation -- again, this was not very visible in the omni-search).
There is however a difference in the handling of column numbers, since in a
lot of cases we now need to increase the finish column number by 1 when
highlighting in the editor.

M906-010

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@229305 936e1b1b-40f2-da11-902a-00137254ae57