Fix /Edit/Paste Previous when using 'auto-indent after paste'
Since the automatic indentation changes the position of the cursor,
paste previous was no longer working. So we need to save the new
position of the cursor instead of computing it. Also, we cannot look
at the length of the paste text, since auto-indent means that more
characters are in fact inserted in some cases.

N709-005

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@230543 936e1b1b-40f2-da11-902a-00137254ae57