Fix timeouts that can be observed when using the remote mode.
The timeouts was due to a protective check making sure that the buffer
was always analysed by the actual call to GNAT.Expect, even when the
underlying process died. The drawback being that with the remote process
dead, expect had to wait the full timeout before exiting if no match
was found in the remaining buffer.

TN: N716-014


git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@231532 936e1b1b-40f2-da11-902a-00137254ae57