Fix errors in tests
The previous patch changed the way the menus are created. As a result,
the first window in gtk_window_list_toplevels is not necessarily a
realized window, but the code for "GPS.send_key_event" was fragile and
was always sending the event to that first window. Instead, we now
search for the first realized window.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gps@231915 936e1b1b-40f2-da11-902a-00137254ae57