Rewrite Entities Browser using Gtkada.Canvas_View
N513-005

This commit removes the "Show Source" contextual menu in the browser,
and replaces it with the standard "Goto declaration" and "Goto body"
menus, which are now forced in this browser.

This commit also removes the contextual menu "Hide/Show inherited ops".
There are better ways to find that information nowadays (completion,
methods.py,...)

Change-Id: I8c775810378d6d3c2bbdcb0f0c330263eef4e1d9