Rewrite the Debugger Data window with Gtkada.Canvas_View
The "auto-refresh" icon was removed from the title bar of the items,
both because this is accessibly from the contextual menu already and
we do not need to make it visible at all times, and because the drawing
of the icon always appears as blurry, even when using an svg image.

The "close" button was replaced with text "X" also to make it sharper.

For now at least, the label on links has been changed: instead of
showing for instance "@(1).all" we show the real name "Var(1).all".
This could result in longer labels, but they are also more readable.

N513-005

Change-Id: I7771e0245b8bf52e053a89ea83b8a7c43873825b