When an item is removed, also remove its aliases
Those aliases were hidden (and extra links set up to point to the item
instead of the alias). So when the user closes the item, he also means
to close the aliases.
This was the behavior implemented when using GtkAda.Canvas, which has
now been ported to Gtkada.Canvas_View

N513-005

Change-Id: I13c072f140f416df83015cf05fe3dd390d5249d0