Fix the "in browser" filter
Pressing '-' or '=' in the omni-search, while the browser had the
focus, would still zoom in or out.
The issue was general to all Gtkada.Canvas_View-based browsers, in
that the focus_widget for the MDI child was set incorrect and in
fact did not have the Can_Focus flag set. As a result, gtk+ did not
transfer the keyboard focus properly, and thus the computation of
the current context was wrong.
N910-027

Change-Id: Ie560f038b007ede982e80fd84140114e28efb983