Add the File Operations menu to the Project view
Immediately open the newly created file, which seems to be the workflow
that people expect.
Remove the warning dialog when removing a file, which says we might have
to modify the project. This interfers with the workflow, might be
inaccurate, and the builder will report an actual error anyway if
needed.

N916-003

Change-Id: If4df4653c155b9480dca037519e37433e43d779c