Handle command line locally
The command line switches are now parsed as part of the applications
local_command_line callback. As a result, this is done before the
application tries to connect to the GUI server, and means we can use
--help and --version even when no graphical environment exists.

Similar to the previous patch, but fixes handling of -Pname (since
Goption does not handle attached parameters, they need a '=' or
space to separate them)

This patch also improves the handling of -X, which is now displayed
along with the other switches in -help.

We now do the initialization of environment variables and traces
earlier (in local_command_line), since otherwise the switch --traceoff
has no effect, resulting in regressions in the testsuite. However, this
means we use our own way to get the current path and environment
variables instead of using gtk+, but that's what we were doing in
previous versions of GPS.

NA22-010

Change-Id: I6d5023b07d99f53191b30822567f16098b84d596