Fix error when closing and expanding a file node.
Load gps.gpr, expand 'src' and 'gps-main.adb', then close 'src' and
open again. This would result in a gtk+ warning, because gtk+ is doing
the node expansion asynchronously but we were changing the children of
the file node to refresh the contents.
Instead, we are now removing the contents when collapsing the file node.

No known-problem entry needed, this was done during the 6.1 development
when adding filters to the project view.

NB03-039

Change-Id: I9507f13fcbc71466945051d44d685408e789592e