Fix project associated with new editors
When a new file is created and then saved, GPS automatically recomputes
the project view, but does not update the project name stored in the
editor. As a result, when the file needs to be opened later for any
reason (for instance going from spec to body and back), this existing
editor is not reused because it does not have the right project.

(Find_Editor): add an extra safety for the above scenario: when an
editor has no project associated, it is always considered a valid
candidate, whatever project parameter was passed in parameter.

(Filename_Changed_Handler): also updates the project names in the
tooltips.

NB05-034

Change-Id: If6a447a8ce086235fa73499a1c7e910c827ec121