Fix the counterpart problem in libclang
The main change in this commit is that the outline does not rely on the
Counterpart function anymore to group spec and bodies together, it will
instead be done by conflating the unique_id (that does not include
wether the element is spec or body, so same unique_id for spec and body)
with the result from the Is_Declaration method.

This was necessary because in C there can be several specs for one body,
and libclang didn't allow to jump from one to the other
straightforwardly.

Also, types for fields were added in the outline in C/C++.

Change-Id: Ia0c09947d89909500c8bccaf1c9ecd344afd0db9