Make sure floating windows are only reset on current monitor.
On some systems, this will be a minor regression since GPS would be able
to open on another monitor, but on other systems, for instance OSX,
having the window on another monitor means it is invisible and can't be
manipulated by the user.
So the least-bad solution is to always restore to current monitor.

NB27-029

Change-Id: Ia2f4ad3f6ffdc0695993569c5aac44046bec7f2e
(cherry picked from commit 8b39511c41ac7c3fc1566dc28637cc6aa9f03554)