Obey the gtk-dialogs-use-header settings
This fixes inconsistencies in the way dialogs are displays. They now all
display action buttons either in the header bar or in the action area at
the bottom, depending on the setting.

(*) the preferences dialog does not work yet, because of the way it is
created.

NB27-010

Change-Id: Idbb1b3efaf5fbfcef0d91d5b978009721d153bae
(cherry picked from commit 51d83dbfaaba0d2d420a3b05317dda4a3240d07f)