New plug-in: gnatemulator.py to run and debug on GNATemu
The only known limitation right now is the handling of boards. It's
possible for the user to specify the board in the project file, but we
could improve the plug-in by guessing board name from run-time name.

This patch also removes GNATemu handling from board_support plug-in, 
and renames it to st_link.py because it now only contains workflows for
st-link tools.

Part of NB27-008.

Change-Id: If77c93c71d75bcc5d7210ed3f904d344edbc76ad