Update invocations of the GNAT driver to work with the Target attribute
Now GPS handles projects' Target attribute, it should consider it when
present in order to choose what process to spawn for GNATtest,
GNATmetrics, etc. (i.e. what GNAT driver to spawn) while still
supporting GNAT* attributes from the IDE package.

This change adapts various plug-ins in order to consider this attribute.
In order to ease this, it also introduces a new "%gnat" macro for
command line expansion: this will expand to the correct command for the
GNAT driver according to the Target or the IDE attributes.

Fix for NC09-033

Change-Id: Iafb84574dafeb1f1cdf330cf3c92a07a5a03b1c8