Improve transition from the old emacs.py plugin to the Key Shortcut theme
Users of GPS 6.0.2 or before might have activated the emacs.py plug-in to
get the Emacs key bindings. This is however done via the Key Shortcuts
dialog in GPS 6.1.0, and this check in improves the transition from one
setup to the other.
It provides a dummy emacs.py plugin which disables itself for the next
restart, and activated the emacs key shortcut.
However, users that will then launch older versions of GPS again will no
longer have the Emas keybindings as a result.

("GPS.History"): new python class

NB11-004

Change-Id: I78171f9d002558dcb1316fad592c1628faec44d8