Cleanup the elaboration of the GNATcoverage integration plug-in
Move part of the elaboration in the "gps_started" hook, as required for
menus creation (for custom targets) and leave the rest before this hook,
as required for the "gnatcov" build mode. Also fix PEP8 issues switching
XML string literals to the dedicated XML generation helpers (and improve
them as needed).

Fix for NC11-029

Change-Id: Ief5714fadd09d8ab53c6ea007c7f2a06ae516ed9