Review exit sequence
Cleanup the exit sequence now that we are using GtkApplication a bit more.
Some cleanup needs to be done when the window is closed, before the
application shuts down. The latter should be reserved to cleanup
GUI-independent data.
The kernel again stores a pointer to the main_window, rather than its id
in the application. The latter did not work well on exit, since the id
would no longer be valid by the time the main window's children are
destroyed. As a result, plugins like filepos.py no longer had access
to the mdi (through the main window) and could not save the position.
The main window no longer stores the kernel, which instead is available
via the application. Otherwise we could no longer get access to the kernel
as part of the application shutdown.

Fix the handling of the delete_event for the main window. Its goal is not
to actually exit the application, only to indicate whether it is OK to do
so. This gives better control on the exit sequence, too.

Connect the child_selected signal later in the startup sequence. This fixes
an issue where we were trying to emit the context_changed signal before it
been registered (when creating the Messages view). This also speeds up the
startup a bit, since we are handling fewer context_changed hook, and
therefore creating fewer contexts.

Remove some global variables in gps-callbacks.ads, which could be replaced
by better API now.

NC11-021

Change-Id: I3ee22cccb81a7c4fa29d2e37914829ecc4e7f805