Review handling of selection contexts
The current context is no longer built on demand. Instead, the view are
responsible for creating the new context whenever it makes sense (for
instance the selection changed), and pass it to the Context_Changed
signal. This lets views decide when a new context should be computed,
rather than let the caller decide.
In particular, editors will be able to update the context only when
the cursor location has changed (and after a short delay), rather than
for every button press signal (to optimize NB19-029).
Rename and simplify the profile for a number of subprograms related to
contexts, in GPS.Kernel.Modules.UI. In particular, the context creation
is always done via a new primitive operation of GPS_MDI_Child_Record,
and the context_func passed to Instrument_Contextual_Menu is now only
used to add entries to the menu, rather than also compute the context.
This context_func will be removed as soon as all contextual menus are
implemented via GPS actions or in the local config menu.

NB19-029 (2/2)

Change-Id: Ib68a848c62bbc20563e95f5a037e1b388a24ed51