Split /Build/Project, /Build/Run and /Debug/Initialize by project
These menus could become large when the project has lots of main, and
thus become hard to manage.
This required a cleanup in the handling of underscore characters in menu
items, since functions like Register_Menu had a Use_Mnemonics parameter
that only applied to the child menu, not to its parent. As a result,
a menu like /Build/Project/Gnatcoll_Tools/gnatcoll_db2ada ended up not
having underscores displayed properly.

The Build contextual menu on projects (in particular in the Project view)
now only proposes the main units for the current project, not for all
projects.

HB23-004

Change-Id: I312cbfd8812581a565a460befcaa573a95dd7ab7