Reorganize the Project Properties dialog
We no longer use a notebook (which might contain other notebooks for
instance in the switches or naming pages). Instead, we use a tree, as
we do in the preferences dialog.
Each page for the switches has its own entry in the tree. However,
wizards and the file setting editor still use a notebook for the
switches.
Rename and reorganize some of the pages in the Project Properties
dialog (splitting the largest pages when possible).

Move the Toolchain settings to a separate page in the Project Properties
Code moved from toolchains/ui to prj_viewer to avoid cyclic
dependency.
Disabling compilation for a language is no longer done in the list of
languages, but in the Toolchains page via a new checkbox.
When a language has an invalid compiler, the name of the language is no
longer highlighted in red in the list of languages.

O119-023 (1/2)

Change-Id: I6dd90f2f000bdf4fa1ab5093ff8a9e93643d55bd