Stabilize exit sequence when debugger is open
- When we are about to return True from the callback to
delete_event on the main window, set the kernel destruction
flag. This is a safeguard to make sure we do not update
menus after the point of destruction of the main window.

- In reaction to the "before_exit_action_hook", close the
debuggers. This is needed so that we are able to save the
debug perspective and restore the default perspective
before the main window is destroyed.

Prevents segmentation faults when exiting GPS while
a debugger is open.

For O122-011.

Change-Id: I1b00def37ee1ad115f02a1ec37822d7b06f08f7e