Fix for O120-033: Fix Is_In_Comment for Ada.
Make sure that Is_In_Comment has up to date information, at least for
Ada, since it is used for auto casing synchronously.

As discussed with Manu, and per his comments on the review, the fix
to Is_In_Comment is limited to the Autocase_Text function, so as to 
limit the scope of the hack, since it is a non optimal solution.

Change-Id: I09d329ee04debd8ebf09969f43ad88184e005846