Fix display of debugger toolbar buttons.
When loading a new project, these buttons became visible and sensitive
again.
As a security, buttons are now always set insensitive even if they will
be hidden.
We also do not automatically force a showing of the buttons as part of
a call to Show_All (not sure which part of the code is doing that
show_all upon loading a project).
Tentatively remove code that was refreshing the iconname for a button.
This does not seem to be needed anymore now that we have moved away
from the stock icons.

Fixes O211-048

Change-Id: I0a958a8b4d920fc876d112659068aa2951e9788e