Avoid calls to the loop that recompute the menu visibility, during startup
Every time we register an action, we were starting the loop, which did at
least one iteration of computing the filters.
Also start monitoring the "context_changed" hook once we have finished
loading all the plug-ins, also to avoid calls to this loop.
Remove one duplicate call to this loop when registering a new menu.

Part of O213-025

Change-Id: I023906cab30fac9620e9a68d26ffb82c6b0bb0f3