Quick fix for missing title in assistant window.
This fixes regressions in J914_009 after the move to 3.14.
The title should now be associated with each page, and is automatically
set by the assistant when the page changes. Unfortunately, when we do
this, the assistant also adds another tree on the left side for the
various pages, so the code in this package will likely need more
refactoring.

Change-Id: Icd31be31f0c333d18206839a9781c466097f3beb