(GPS_Dialog): new type
All dialogs in GPS should either be a full MDI_Child, or derived from
GPS_Dialog. The goal is to keep the context up-to-date when the dialog
gets the focus.
Fixes an issue in particular with the Codepeer Message Review dialog,
where pressing Backspace would in fact delete characters in the editor
that previously had the focus. Also changes the way the dialog is
created so that the action buttons are put in the proper location based
on the gtk settings, for consistency with the rest of GPS.

O209-022

Change-Id: Ic28a6121305426ff3a8aba4e743859edc889a136