("GPS.Action.disable"): new python function
Make the Action_Record type private, and provide accessors for its fields
instead.
This provides a single way to check whether an action filter matches, and
let's the user fully disable a command. All associated menus and buttons
are also grayed out.

This is used in particular for the prevent_project_editor.py plugin

O327-030

Change-Id: I7b1b7278aa04c0aa2fcdaf419efde0cc61aaae80