Replace history keys with standard preferences for Project view
Provide an API for preferences similar to the one we have for histories
(associating with a widget and have two-way synchronization). This makes
it easier to use preferences to configure views.
In particular, it is possible to use the python API to configure the
project view, as requested for O326-046

Change-Id: I5cfe980392560385a772f8fcad60045f615b9359