Do not use so many colors for auto_highlight_occurrences
Instead, define two colors that are going to be reused
across plugins.

Modify textmate.py to take this into account; also refine
that plugin.

This also fixes an issue where a change in the preference
colors did not apply immediately to the selected items
in the text.

Allow users to specify transparent foregrounds for style
variants: this allows specifying variants which only modify
the background.

For O501-009 (1/2)

Change-Id: Ie0d6e9534c5ea5bbdb2bd2269712f4973ccc61c8