clean up handling of st-util
Now st-util related button enabling includes searching project
dependencies and stops when one is found. This allows a project
to extend another one, for example, in which the one extended
has the required attributes.

Also now debug workflow does not issue an explicit "load"
command. It is redundant and causes two loads to occur.

Change-Id: Id0fdb9763d1ce44f249ec44d018c7326bf5eebe6