Only find "Makefile" by default, not "makefile"
Due to the way the build targets are setup, we can only have one default
for the attribute, which is set to "Makefile". So GPS was parsing makefile
to find the possible targets, but then spawning "make -f Makefile", thus
failing on linux systems. Unless, of course, the Make'Makefile attribute
was defined in the project.

O604-006

Change-Id: I001ef0ea105932b553fe71618620f1da4fb03c37