Add support for --config, --configdb and --autoconf
GPS can now load a config file when it parses the project. This is
in particular useful to get information about the naming schemes for
multiple languages, so that users do not have to duplicate that
information for gprbuild, GPS and other gpr-based tools.
This config file is also passed to gnatinspect automatically.

MC03-007

Change-Id: I65f92780b1379e6a6cc65752d3efaa48cb81abdc