Do not open an editor for the "Current File" search context
When no editor is opened yet and omni-search is used, it would open an
"Unnamed" editor, which loses the focus in the omni-search and thus stops
the search (later contexts are not searched, as happend in M625_006)

When searching in source files, it was possible that the search never
ended. This became visible when searching in "current file", since later
search providers were never called (regression in M625_006)

Minor cleanups in search-sources contexts.

for O820-013

Change-Id: I73f5a3395828833ba9ed3e6aa3d1c62ff360e94f