O903-013 Fix column number in SPARK 2014 plugin messages when < 10
Comments from David Hauzar below.

I am attaching a patch that solves the following issue: if the column
number associated to gnatprove message is less than 10, extra
information about this message are not displayed in GPS.

The reason for this is that the text returned by the function
GNATprove_Parser.build_msg_full_text must be exactly the same as the
text returned by gnatprove - in the SPARK gps plugin, this text is used
to match the message produced by gnatprove and extra information about
this message stored in *.spark file (trace, counterexample).

More precisely, using information in the *.spark file, the plugin first
creates a map from the text produced by build_msg_full_text to message
id. And then, when going through messages produced by gnatprove, it uses
the text of the message and this map to get the id of the message. If
these text do not match, id is not found and no extra information about
the message are displayed.

Change-Id: I8ac414543e3cf5fd9e0a6858c153009896619a0d