Fix handling of hooks returning a value when implemented as procedure
Some hooks are setup so that their function returns a value (for instance
the open_file_action_hook excepts a boolean from its callbacks), but running
the hook does not return that final value to the Ada caller. In this case,
we were ignoring the value returned by the python callbacks, and always
running all callbacks.
As a result, we could not override open_file_action_hooks from python, since
the standard action (opening an editor) was always executed any way.

Change-Id: I3adb563d158bd96e4ef61380a439bf159ca53aad