Ensure that the tooltips are always drawn on the screen.
This is a continuation work for the previous commit. This version
takes into account multi-monitors setup. We make sure that when
using multiple physical screens the tooltips are always drawn on
the current screen (that is, where GPS is displayed). And that
tooltips are never drawn outside of this physical screen.

Continued work for O923-032.

Change-Id: Ie942971f6cd04188eaa3b54d648288a4625f542d