Solidify the backend for relaunching the cross-references
Relying on 'on_compilation_finished' is too fragile and seems
broken, since it is not called for 'Recompute Xref Info'.

This was broken in the following scenario:
   - a project is loaded: launch gnatinspect
   - user launches gprbuild while gnatinspect is still runnning
    -> we do not relaunch gnatinspect while one is running,
       but make a note to launch it later
   - 'on_compilation_finished' isn't called
    => the second launch of gnatinspect, which reads the
       value of the .ali files, never happens.

Instead of relying on 'compilation_finished', rely on the
registered OutputParser, and add a method 'gnatinspect_completed'
to handle the logic of the relaunch.

For O923-012.

Change-Id: I869d768d7a96e0b7044943d4e7d0ad68eac01b06