O923-008 Add user profile pref in SPARK with different proof panels
A new preference is defined in the SPARK 2014 plugin, to define the user
profile as either 'Basic' or 'Advanced'. When the user profile is basic,
the proof panel shows on proof actions is simple. When the user profile
is advanced, the proof panel is more complex. In particular, the new --level
switch is only shown in the basic proof panel.

Change-Id: I80e97764447d92decebdc305365ae414fbcd1aaa