Add groups to preferences pages
OA16-057

GPS preferences dialog can now contain groups.
The association between a preference and its group is done via
the 'Page' parameter of the Default_Preferences.Create functions,
by specifying a group name after the preferences page name.

The page names and the group names are delimitered with a ':'
character (eg:"General:Appearance", where "General" is the page name
and "Appearance" the group name).

Note that this is optional: if no group name is given for a page,
the page will contain a default group without any label.

Change-Id: Iab6300e6005e27a7cf803956d473c07984c1fb1c