Fix computation of the list of mains for aggregate projects
The menu /Build/Project/* did not include some of the main units
when the same source file was used by multiple aggregated projects.
Due to limitations in gprbuild, though, there is currently no way
to build a specific main unit for a specific project, so this
issue is not fully solved yet.

The menu /Debug/Initialize/* was not taking the project into account,
so would always debug i1.gpr:i (for instance) even if there is another
project i2.gpr with also an executable named 'i'.
This menu is now computed with the same algorithm as used for /Build,
so that we get the same list of mains and in the same order.

O820-024 (1/2)

Change-Id: Iff5bd55c03f89294fa6b9f04d6b27c7acd05b9bb